Moved the rest client creation and zipping of imported file to module level utils
